// admin/src/app/@core/models/pequeno-grupo.model.ts
export class PequenoGrupoModel {
  id: number;
  congregacaoId: number;
  nome: string;
  liderPequenoGrupoId: number;
  coLiderPequenoGrupoId: number;
  diaSemana: string;
  horarioReuniao: string;
  cep: string;
  endereco: string;
  numero: string;
  complemento: string;
  bairro: string;
  cidade: string;
  estado: string;
  status: string;
  dataCadastro: Date;
}